requirements.txt
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0          # for SQL Server; swap for pymssql if preferred
langchain>=0.2.0
langchain-anthropic>=1.4.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
